Add files via upload
The proposed project aims to develop a Sorting Algorithm Visualizer in Java for the Algorithm Lab. Sorting algorithms play a crucial role in computer science, and understanding their behavior is fundamental for aspiring computer scientists. The Sorting Algorithm Visualizer will provide an interactive platform for students to visualize various sorting algorithms in action, fostering a deeper understanding of their efficiency and performance characteristics.